refactor(outputs): moved output appending to redux
