+++ src/notebook/constants.js
@@ -33,6 +33,7 @@ export const SPLIT_CELL = 'SPLIT_CELL';
 export const ABORT_EXECUTION = 'ABORT_EXECUTION';
 
 export const UPDATE_CELL_EXECUTION_COUNT = 'UPDATE_CELL_EXECUTION_COUNT';
+export const APPEND_OUTPUT = 'APPEND_OUTPUT';
 export const CLEAR_CELL_OUTPUT = 'CLEAR_CELL_OUTPUT';
 
 export const CHANGE_OUTPUT_VISIBILITY = 'CHANGE_OUTPUT_VISIBILITY';
